- Adds an "Authorization" header with the access token to web requests. Some service providers require authorization through this header. - Removes trailing slash from "redirect_url" parameter during authentication. Some service providers break due to this. - Changed toast message upon successful authentication to be more user-friendly. Also made it display for longer so that user has time to read it. - Added Immutable Pending Intent Flag as required by Android 12 (API Level 31).

+++ source/Core/Xamarin.Auth.XamarinAndroid/UINativeNonIntegratedBrowsers/WebAuthenticatorNativeBrowserActivity.cs
@@ -282,7 +282,7 @@ namespace Xamarin.Auth._MobileServices
 
                             if (msg != null)
                             {
-                                Toast.MakeText(this, msg, ToastLength.Short).Show();
+                                Toast.MakeText(this, msg, ToastLength.Long).Show();
                             }
                         }
                     );
